requirements.txt
google-generativeai>=0.8.0
google-adk>=1.0.0
arxiv>=2.1.0
duckduckgo-search>=6.0.0
ddgs>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.30.0
